requirements.txt.txt
python-telegram-bot==20.3
gspread
pandas
google-auth
